test(list): combine character-counter screeneshot
